+++ actors.h
@@ -13,91 +13,91 @@ class Actor {
     Actor();
 
     bool has_shadow();
-    float get_reflect();
+    double get_reflect();
 
     virtual ~Actor();
-    virtual float solve(Eigen::Vector3f *origin, Eigen::Vector3f *direction,
-                        float mind, float maxd) = 0;
+    virtual double solve(Eigen::Vector3d *origin, Eigen::Vector3d *direction,
+                        double mind, double maxd) = 0;
 
-    virtual Pixel pick_pixel(Eigen::Vector3f *hit,
-                             Eigen::Vector3f *normal) = 0;
+    virtual Pixel pick_pixel(Eigen::Vector3d *hit,
+                             Eigen::Vector3d *normal) = 0;
 
-    virtual Eigen::Vector3f calculate_normal(Eigen::Vector3f *hit) = 0;
+    virtual Eigen::Vector3d calculate_normal(Eigen::Vector3d *hit) = 0;
 
   protected:
     bool has_shadow_;
-    float reflect_;
+    double reflect_;
     Texture *texture_;
 };
 
 
 class Plane : public Actor {
   public:
-    Plane(Eigen::Vector3f *center, Eigen::Vector3f *normal, float scale, 
-          float reflect, const char *texture);
+    Plane(Eigen::Vector3d *center, Eigen::Vector3d *normal, double scale, 
+          double reflect, const char *texture);
 
     virtual ~Plane();
-    virtual float solve(Eigen::Vector3f *origin, Eigen::Vector3f *direction, 
-                        float mind, float maxd) override;
+    virtual double solve(Eigen::Vector3d *origin, Eigen::Vector3d *direction, 
+                        double mind, double maxd) override;
    
-    virtual Pixel pick_pixel(Eigen::Vector3f *hit,
-                             Eigen::Vector3f *normal) override;
+    virtual Pixel pick_pixel(Eigen::Vector3d *hit,
+                             Eigen::Vector3d *normal) override;
     
-    virtual Eigen::Vector3f calculate_normal(Eigen::Vector3f *hit) override;
+    virtual Eigen::Vector3d calculate_normal(Eigen::Vector3d *hit) override;
 
   private:
-    Eigen::Vector3f center_;
-    Eigen::Vector3f normal_;
-    Eigen::Vector3f tx_;
-    Eigen::Vector3f ty_;
-    float scale_;
+    Eigen::Vector3d center_;
+    Eigen::Vector3d normal_;
+    Eigen::Vector3d tx_;
+    Eigen::Vector3d ty_;
+    double scale_;
 };
 
 
 class Sphere : public Actor {
   public:
-    Sphere(Eigen::Vector3f *center, float radius, Eigen::Vector3f *axis, 
-           float reflect, const char *texture);
+    Sphere(Eigen::Vector3d *center, double radius, Eigen::Vector3d *axis, 
+           double reflect, const char *texture);
 
     virtual ~Sphere();
-    virtual float solve(Eigen::Vector3f *origin, Eigen::Vector3f *direction, 
-                        float mind, float maxd) override;
+    virtual double solve(Eigen::Vector3d *origin, Eigen::Vector3d *direction, 
+                        double mind, double maxd) override;
 
-    virtual Pixel pick_pixel(Eigen::Vector3f *hit,
-                             Eigen::Vector3f *normal) override;
+    virtual Pixel pick_pixel(Eigen::Vector3d *hit,
+                             Eigen::Vector3d *normal) override;
 
-    virtual Eigen::Vector3f calculate_normal(Eigen::Vector3f *hit) override;
+    virtual Eigen::Vector3d calculate_normal(Eigen::Vector3d *hit) override;
 
   private:
-    Eigen::Vector3f center_;
-    Eigen::Vector3f tx_;
-    Eigen::Vector3f ty_;
-    Eigen::Vector3f tz_;
-    float R_;
+    Eigen::Vector3d center_;
+    Eigen::Vector3d tx_;
+    Eigen::Vector3d ty_;
+    Eigen::Vector3d tz_;
+    double R_;
 };
 
 
 class Cylinder : public Actor {
   public:
-    Cylinder(Eigen::Vector3f *center, Eigen::Vector3f *direction, 
-             float radius, float span, float reflect, const char *texture);
+    Cylinder(Eigen::Vector3d *center, Eigen::Vector3d *direction, 
+             double radius, double span, double reflect, const char *texture);
 
     virtual ~Cylinder();
-    virtual float solve(Eigen::Vector3f *origin, Eigen::Vector3f *direction, 
-                        float mind, float maxd) override;
+    virtual double solve(Eigen::Vector3d *origin, Eigen::Vector3d *direction, 
+                        double mind, double maxd) override;
 
-    virtual Pixel pick_pixel(Eigen::Vector3f *hit,
-                             Eigen::Vector3f *normal) override;
+    virtual Pixel pick_pixel(Eigen::Vector3d *hit,
+                             Eigen::Vector3d *normal) override;
 
-    virtual Eigen::Vector3f calculate_normal(Eigen::Vector3f *hit) override;
+    virtual Eigen::Vector3d calculate_normal(Eigen::Vector3d *hit) override;
 
   private:
-    Eigen::Vector3f A_;
-    Eigen::Vector3f B_;
-    Eigen::Vector3f tx_;
-    Eigen::Vector3f ty_;
-    float R_;
-    float span_;
+    Eigen::Vector3d A_;
+    Eigen::Vector3d B_;
+    Eigen::Vector3d tx_;
+    Eigen::Vector3d ty_;
+    double R_;
+    double span_;
 };
 
 }  // namespace mrtp
