+++ renderer.cpp
@@ -12,8 +12,8 @@
 
 namespace mrtp {
 
-static const float kDegreeToRadian = M_PI / 180.0f;
-static const float kRealToByte = 255.0f;
+static const double kDegreeToRadian = M_PI / 180;
+static const double kRealToByte = 255;
 
 /*
 distance: a distance to fully darken the light
@@ -22,8 +22,8 @@ bias: correction to shadows to avoid self-intersection
 distance: maximum distance reached by rays
 maxdepth: number of recursion levels of a reflected ray
 */
-Renderer::Renderer(World *world, int width, int height, float fov,
-                   float distance, float shadow, float bias, int maxdepth,
+Renderer::Renderer(World *world, int width, int height, double fov,
+                   double distance, double shadow, double bias, int maxdepth,
                    int nthreads, const char *path) : 
     world_(world), 
     width_(width), 
@@ -36,8 +36,8 @@ Renderer::Renderer(World *world, int width, int height, float fov,
     nthreads_(nthreads), 
     path_(path) {
 
-    ratio_ = static_cast<float>(width_) / static_cast<float>(height_);
-    perspective_ = ratio_ / (2.0f * std::tan(kDegreeToRadian * fov_ / 2.0f));
+    ratio_ = static_cast<double>(width_) / static_cast<double>(height_);
+    perspective_ = ratio_ / (2 * std::tan(kDegreeToRadian * fov_ / 2));
 
     Pixel dummy;
     framebuffer_.assign(width_ * height_, dummy);
@@ -63,8 +63,8 @@ bool Renderer::write_scene() {
     return rs_ok;
 }
 
-bool Renderer::solve_shadows(Eigen::Vector3f *origin, Eigen::Vector3f *direction,
-                             float maxdist) {
+bool Renderer::solve_shadows(Eigen::Vector3d *origin, Eigen::Vector3d *direction,
+                             double maxdist) {
     std::vector<Actor *> actors = world_->ptr_actors_;
     std::vector<Actor *>::iterator iter = actors.begin();
     std::vector<Actor *>::iterator iter_end = actors.end();
@@ -72,8 +72,8 @@ bool Renderer::solve_shadows(Eigen::Vector3f *origin, Eigen::Vector3f *direction
     for (; iter != iter_end; ++iter) {
         Actor *actor = *iter;
         if (actor->has_shadow()) {
-            float distance = actor->solve(origin, direction, 0.0f, maxdist);
-            if (distance > 0.0f) {
+            double distance = actor->solve(origin, direction, 0.0, maxdist);
+            if (distance > 0.0) {
                 return true;
             }
         }
@@ -81,8 +81,8 @@ bool Renderer::solve_shadows(Eigen::Vector3f *origin, Eigen::Vector3f *direction
     return false;
 }
 
-Actor *Renderer::solve_hits(Eigen::Vector3f *origin, Eigen::Vector3f *direction,
-                            float *currd) {
+Actor *Renderer::solve_hits(Eigen::Vector3d *origin, Eigen::Vector3d *direction,
+                            double *currd) {
     std::vector<Actor *> actors = world_->ptr_actors_;
     std::vector<Actor *>::iterator iter = actors.begin();
     std::vector<Actor *>::iterator iter_end = actors.end();
@@ -90,8 +90,8 @@ Actor *Renderer::solve_hits(Eigen::Vector3f *origin, Eigen::Vector3f *direction,
 
     for (; iter != iter_end; ++iter) {
         Actor *actor = *iter;
-        float distance = actor->solve(origin, direction, 0.0f, maxdist_);
-        if ((distance > 0.0f) && (distance < (*currd))) {
+        double distance = actor->solve(origin, direction, 0.0, maxdist_);
+        if ((distance > 0.0) && (distance < (*currd))) {
             *currd = distance;
             hit = actor;
         }
@@ -99,49 +99,49 @@ Actor *Renderer::solve_hits(Eigen::Vector3f *origin, Eigen::Vector3f *direction,
     return hit;
 }
 
-Pixel Renderer::trace_ray_r(Eigen::Vector3f *origin, Eigen::Vector3f *direction, int depth) {
+Pixel Renderer::trace_ray_r(Eigen::Vector3d *origin, Eigen::Vector3d *direction, int depth) {
     Pixel pixel;
-    pixel << 0.0f, 0.0f, 0.0f;
+    pixel << 0, 0, 0;
 
-    float currd = maxdist_;
+    double currd = maxdist_;
     Actor *hitactor = solve_hits(origin, direction, &currd);
 
     if (hitactor) {
-        Eigen::Vector3f inter = ((*direction) * currd) + (*origin);
-        Eigen::Vector3f normal = hitactor->calculate_normal(&inter);
+        Eigen::Vector3d inter = ((*direction) * currd) + (*origin);
+        Eigen::Vector3d normal = hitactor->calculate_normal(&inter);
 
         // Calculate light intensity
-        Eigen::Vector3f tolight = world_->ptr_light_->calculate_ray(&inter);
+        Eigen::Vector3d tolight = world_->ptr_light_->calculate_ray(&inter);
 
-        float lightd = tolight.norm();
-        tolight *= (1.0f / lightd);
+        double lightd = tolight.norm();
+        tolight *= (1.0 / lightd);
 
-        float intensity = tolight.dot(normal);
+        double intensity = tolight.dot(normal);
 
-        if (intensity > 0.0f) {
+        if (intensity > 0.0) {
             // Prevent self-intersection
-            Eigen::Vector3f corr = inter + bias_ * normal;
+            Eigen::Vector3d corr = inter + bias_ * normal;
 
             // Check if the intersection is in a shadow
             bool isshadow = solve_shadows(&corr, &tolight, lightd);
-            float shadow = (isshadow) ? shadow_ : 1.0f;
+            double shadow = (isshadow) ? shadow_ : 1.0;
 
             // Decrease light intensity for actors away from the light
-            float ambient = 1.0f - std::pow(lightd / maxdist_, 2);
+            double ambient = 1.0 - std::pow(lightd / maxdist_, 2);
 
             // Combine pixels
-            float lambda = intensity * shadow * ambient;
+            double lambda = intensity * shadow * ambient;
 
             Pixel pick = hitactor->pick_pixel(&inter, &normal);
-            pixel = (1.0f - lambda) * pixel + lambda * pick;
+            pixel = (1.0 - lambda) * pixel + lambda * pick;
 
             // If the hit actor is reflective, trace a reflected ray
             if (depth < maxdepth_) {
-                float coeff = hitactor->get_reflect();
-                if (coeff > 0.0f) {
-                    Eigen::Vector3f ray = (*direction) - (2.0f * direction->dot(normal)) * normal;
+                double coeff = hitactor->get_reflect();
+                if (coeff > 0.0) {
+                    Eigen::Vector3d ray = (*direction) - (2.0 * direction->dot(normal)) * normal;
                     Pixel reflected = trace_ray_r(&corr, &ray, depth + 1);
-                    pixel = (1.0f - coeff) * reflected + coeff * pixel;
+                    pixel = (1.0 - coeff) * reflected + coeff * pixel;
                 }
             }
         }
@@ -154,8 +154,8 @@ void Renderer::render_block(int block, int nlines) {
 
     for (int j = 0; j < nlines; j++) {
         for (int i = 0; i < width_; i++, pixel++) {
-            Eigen::Vector3f origin = world_->ptr_camera_->calculate_origin(i, j + block * nlines);
-            Eigen::Vector3f direction = world_->ptr_camera_->calculate_direction(&origin);
+            Eigen::Vector3d origin = world_->ptr_camera_->calculate_origin(i, j + block * nlines);
+            Eigen::Vector3d direction = world_->ptr_camera_->calculate_direction(&origin);
             *pixel = trace_ray_r(&origin, &direction, 0);
         }
     }
@@ -173,7 +173,7 @@ If nthreads=0, uses as many threads as available.
 Returns rendering time in seconds, corrected for
 the number of threads.
 */
-float Renderer::render_scene() {
+double Renderer::render_scene() {
     world_->ptr_camera_->calculate_window(width_, height_, perspective_);
 
     int time_start = clock();
@@ -207,9 +207,9 @@ float Renderer::render_scene() {
 #endif //!_OPENMP
 
     int time_stop = std::clock();
-    float time_used = static_cast<float>(time_stop - time_start) / CLOCKS_PER_SEC;
+    double time_used = static_cast<double>(time_stop - time_start) / CLOCKS_PER_SEC;
     if (nthreads_ > 1) {
-        time_used *= 1.0f / static_cast<float>(nthreads_);
+        time_used *= 1 / static_cast<double>(nthreads_);
     }
     return time_used;
 }
